Code Cleanup + Battle Engine Fix
I removed comments and blank lines to clean up the code. Through doing so I noticed a mismatch between the deployed code on glitch and the repo code specifically having to do with the battle engine. I reverted the repo battle engine to this working, cleaned version from the glitch deployment.